refactor[InsertSapiensMinutas]
